toolkit/content/tests/chrome/test_statusbar.xul

<?xml version="1.0"?>
<?xml-stylesheet href="chrome://global/skin" type="text/css"?>
<?xml-stylesheet href="chrome://mochikit/content/tests/SimpleTest/test.css" type="text/css"?>
<!--
  XUL Widget Test for statusbar
  -->
<window title="Statusbar Test"
        onload="setTimeout(test_statusbar, 0);"
        xmlns="http://www.mozilla.org/keymaster/gatekeeper/there.is.only.xul">
  <script type="application/javascript"
          src="chrome://mochikit/content/tests/SimpleTest/SimpleTest.js"></script>      
  <script type="application/javascript"
          src="chrome://mochikit/content/tests/SimpleTest/EventUtils.js"></script>      

<statusbar>
  <statusbarpanel id="panel" label="OK" image="happy.png"/>
</statusbar>

  <!-- test results are displayed in the html:body -->
  <body xmlns="http://www.w3.org/1999/xhtml" style="height: 300px; overflow: auto;"/>

  <!-- test code goes here -->
  <script type="application/javascript"><![CDATA[

SimpleTest.waitForExplicitFinish();

function test_statusbar()
{
  var panel = $("panel");
  ok(panel.label, "OK", "statusbarpanel label");
  ok(panel.image, "happy.png", "statusbarpanel image");
  panel.src = "sad.png";
  ok(panel.src, "sad.png", "statusbarpanel set src");
  ok(panel.getAttribute("src"), "sad.png", "statusbarpanel set src attribute");

  SimpleTest.finish();
}

]]>
</script>

</window>
